refactor(server): Added src folder
