+++ server/src/config/cors.js
