+++ server/src/routes/posts/index.js
